+++ packages/wildness_ui/lib/wildness/wildness_properties.dart
@@ -16,6 +16,32 @@ final class WildnessProperties {
     this.maxScaleFactor = 1.2,
   });
 
+  /// Creates a [WildnessProperties] from [Components] and optional [Resources] for light and dark modes.
+  factory fromTheme({
+    Brightness? forceThemeMode,
+    Components? lightComponents,
+    Components? darkComponents,
+    Resources? lightResources,
+    Resources? darkResources,
+    ScrollPhysics physics = const ClampingScrollPhysics(),
+    double minScaleFactor = 0.5,
+    double maxScaleFactor = 1.2,
+  }) =>
+      WildnessProperties(
+        forceThemeMode: forceThemeMode,
+        components: Configuration.fromComponents(
+          light: lightComponents,
+          dark: darkComponents,
+        ),
+        resources: Configuration.fromResources(
+          light: lightResources,
+          dark: darkResources,
+        ),
+        physics: physics,
+        minScaleFactor: minScaleFactor,
+        maxScaleFactor: maxScaleFactor,
+      );
+
   /// Components of this theme.
   ///
   /// To define components, pass an [Iterable] containing one or more [WildnessBase] and their kinds