+++ packages/wildness_ui/lib/wildness/wildness_provider.dart
@@ -1,13 +1,27 @@
 part of '../wildness.dart';
 
 @immutable
-final class WildnessProvider extends InheritedWidget {
+final class WildnessProvider extends InheritedModel<Type> {
   const new({required this.data, required super.child, super.key});
 
   final Wildness data;
 
   @override
   bool updateShouldNotify(WildnessProvider oldWidget) => data != oldWidget.data;
+
+  @override
+  bool updateShouldNotifyDependent(
+    WildnessProvider oldWidget,
+    Set<Type> dependencies,
+  ) {
+    for (final aspect in dependencies) {
+      if (data.components[aspect] != oldWidget.data.components[aspect] ||
+          data.resources[aspect] != oldWidget.data.resources[aspect]) {
+        return true;
+      }
+    }
+    return false;
+  }
 }
 
 @immutable